third_party/WebKit/LayoutTests/fast/events/focus-event-source-device-from-touch-expected.txt

Test that focus events from a touch device should have a sourceDevice and its firesTouchEvents should be true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


tests on div
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true

tests on input
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true

tests on button
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true

tests on textarea
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true

tests on a
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true

PASS successfullyParsed is true

TEST COMPLETE
